Fix triggers for actions with a "text" attribute.
For actions like `DoScript`, the action itself invokes the script by
retrieving the script contents via `getValueDictByKey`, but we were
never setting the dict key. It's not clear if this is a symptomatic
issue or if the triggers use the text as a fallback if the key is not
set, but this does at least cut down on differences between what pydcs
produces and what the mission editor produces, which makes debugging
easier.

The changes to `DoScript` and `MarkTo*` were determined by diffing a
mission file before and after re-saving the mission from the editor.
The changes to other action types are an assumption.