Add multirole fighters target option.
Not sure when this got added to DCS, but multirole fighters are a
separte target type now.

The task also emits a `value` field in addition to `targetTypes` that is
the same thing but a semicolon delimited list.

The ME also produces a `noTargetTypes` which is like `targetTypes`,
but for all the things that *weren't* selected. I've omitted that since
it doesn't seem to matter.