Fix payload clobbering bug, improve lookup.
There was a bug where the max key was being computed based on the max
key of the *input* dict rather than of the dict being modified, so if
the new loadout had an index equal to the current index + 1, we'd
clobber the existing (potentially non-matching) loadout.

This doesn't actually need to be an integer indexed dict though, as this
is never used directly; it's a data structure internal to the class. As
such, I replaced the max index tracking with a name -> payload dict.
This also improves the name based loadout lookup.