Plumb Terrain into Point.
Transformations from X/Y to lat/lon are map specific, so to offer those
APIs on Point we either need to store a reference to the terrain the
point is associated with or any conversion API would need to pass the
terrain. It's annoying for the caller to have to plumb the terrain
through their project, but it's fairly easy for us to track the terrain
with Point.

The Task subclasses depend on Point being default constructible because
create_from_dict creates the type before replacing its properties, and
create_from_dict must have the same signature for all subtypes of Task.
I've introduced Vector2 to cover the cases primarily to cover the cases
where the value is an offset rather than a position, but have also used
that in Task to work around that bit of ugliness.

https://github.com/pydcs/dcs/issues/17