Add Point API for creating new Points.
A common case I'm finding is that I need to take the X/Y coordinates of
a point, use them in another system (such as shapely), then convert the
result back into a Point. I'm currently doing that by accessing the
(supposedly private) _terrain member. This adds a less invasive method
for doing that.